Move OTLP adapters to otproto module for use outside of gRPC exporters.
