+++ sdk-extensions/otproto/README.md
@@ -2,8 +2,8 @@
 
 [![Javadocs][javadoc-image]][javadoc-url]
 
-This module contains code to helps with conversions from OpenTelemetry proto objects to API or SDK
-objects (e.g. SpanId, TraceId, TraceConfig etc.).
+This module contains code to helps with conversions betewen OpenTelemetry proto objects and API or 
+SDK objects (e.g. SpanId, TraceId, TraceConfig, SpanData etc.).
 
 [javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-sdk-contrib-otproto.svg
 [javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-sdk-contrib-otproto
