+++ sdk-extensions/otproto/build.gradle
@@ -16,6 +16,8 @@ dependencies {
     implementation libraries.protobuf,
             libraries.protobuf_util
 
+    testImplementation project(':opentelemetry-sdk-testing')
+
     testImplementation "io.grpc:grpc-testing:${grpcVersion}"
     testRuntime "io.grpc:grpc-netty-shaded:${grpcVersion}"
