+++ sdk-extensions/otproto/src/main/java/io/opentelemetry/sdk/extension/otproto/CommonProperties.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.exporter.otlp;
+package io.opentelemetry.sdk.extension.otproto;
 
 public class CommonProperties {
   public static final String KEY_TIMEOUT = "otel.exporter.otlp.timeout";
