+++ sdk-extensions/otproto/src/main/java/io/opentelemetry/sdk/extension/otproto/package-info.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/** Utilities for working with the OTLP format for traces. */
+/** Utilities for working with the OTLP format. */
 @ParametersAreNonnullByDefault
 package io.opentelemetry.sdk.extension.otproto;
