+++ sdk-extensions/otproto/src/test/java/io/opentelemetry/sdk/extension/otproto/ResourceAdapterTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.exporter.otlp;
+package io.opentelemetry.sdk.extension.otproto;
 
 import static io.opentelemetry.api.common.AttributeKey.booleanKey;
 import static io.opentelemetry.api.common.AttributeKey.doubleKey;
@@ -17,7 +17,6 @@ import io.opentelemetry.proto.common.v1.KeyValue;
 import io.opentelemetry.sdk.resources.Resource;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for {@link ResourceAdapter}. */
 class ResourceAdapterTest {
   @Test
   void toProtoResource() {
